ip_tracking/requirements.txt
Django>=5.0
django-ipware>=7.0
django-ratelimit>=4.1
requests>=2.32
celery>=5.3
redis>=5.0
